We now use users.json for instances for multiple discord users & update all packages

+++ index.js
@@ -1,20 +1,16 @@
 require("dotenv").config();
-const {
-  DISCORD_PREFIX,
-  DISCORD_TOKEN,
-  AWS_INSTANCE_ID,
-  DISCORD_CHANNEL_ID,
-} = process.env;
+const { DISCORD_PREFIX, DISCORD_TOKEN, DISCORD_CHANNEL_ID } = process.env;
 const Discord = require("discord.js");
 const client = new Discord.Client();
 const ec2 = new (require("@aws-sdk/client-ec2").EC2)();
+const users = require("./users.json");
 
 client.on("ready", () => {
   console.log(`Logged in as ${client.user.tag}!`);
 });
 
 client.on("message", async (msg) => {
-  // Check ifm essage was not from a bot
+  // Check if message was not from a bot
   if (msg.author.bot) return;
 
   // Check if message was sent in guild
@@ -29,50 +25,67 @@ client.on("message", async (msg) => {
   const args = msg.content.slice(DISCORD_PREFIX.length).trim().split(/ +/);
   const command = args.shift().toLowerCase();
 
+  const user = users.find((user) => user.discordId === msg.author.id);
+
   if (command === "status") {
-    const response = await ec2.describeInstances({
-      InstanceIds: [AWS_INSTANCE_ID],
-    });
+    let embed = new Discord.MessageEmbed().setTitle("Status");
 
-    if (response.Reservations) {
-      const instance = response.Reservations[0].Instances[0];
+    for (let i = 0; i < user.instances.length; i++) {
+      const response = await ec2.describeInstances({
+        InstanceIds: [user.instances[i]],
+      });
 
-      return msg.channel.send(
-        new Discord.MessageEmbed()
-          .setTitle("Status")
+      if (response.Reservations) {
+        const instance = response.Reservations[0].Instances[0];
+
+        embed
+          .addField("Instance Id", i)
           .addField("Instance Type", instance.InstanceType)
           .addField("Platform", instance.Platform)
           .addField("Public IP Address", instance.PublicIpAddress)
-          .addField("State", instance.State.Name)
-      );
+          .addField("State", instance.State.Name);
+
+        if (user.instances.length - 1 !== i) embed.addField("\u200B", "\u200B");
+      }
     }
+
+    return msg.channel.send(embed);
   } else if (command === "start") {
+    if (args && !user.instances[args[0]])
+      return msg.channel.send(`No instance with id ${args[0]} exists!`);
+
     const response = await ec2.startInstances({
-      InstanceIds: [AWS_INSTANCE_ID],
+      InstanceIds: [user.instances[args[0]]],
     });
 
     return msg.channel.send(
       new Discord.MessageEmbed()
-        .setTitle("Successfully started instance")
+        .setTitle(`Successfully started instance id ${args[0]}`)
         .addField(
           "Instance State",
           response.StartingInstances[0].CurrentState.Name
         )
     );
   } else if (command === "stop") {
+    if (args && !user.instances[args[0]])
+      return msg.channel.send(`No instance with id ${args[0]} exists!`);
+
     const response = await ec2.stopInstances({
-      InstanceIds: [AWS_INSTANCE_ID],
+      InstanceIds: [user.instances[args[0]]],
     });
 
     return msg.channel.send(
       new Discord.MessageEmbed()
-        .setTitle("Successfully stopped instance")
+        .setTitle(`Successfully stopped instance id ${args[0]}`)
         .addField(
           "Instance State",
           response.StoppingInstances[0].CurrentState.Name
         )
     );
   } else if (command === "terminate") {
+    if (args && !user.instances[args[0]])
+      return msg.channel.send(`No instance with id ${args[0]} exists!`);
+
     const response = await ec2.terminateInstances({
       InstanceIds: [AWS_INSTANCE_ID],
     });
